+++ packages/labs/ssr/src/test/integration/tests/basic.ts
@@ -50,6 +50,10 @@ interface ClickableButton extends HTMLButtonElement {
   __wasClicked: boolean;
   __wasClicked2: boolean;
 }
+interface ClickableInput extends HTMLInputElement {
+  __wasClicked: boolean;
+  __wasClicked2: boolean;
+}
 
 const throwIfRunOnServer = () => {
   if (!(globalThis instanceof window.constructor)) {
@@ -1557,6 +1561,23 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
+  'AttributePart on void element': {
+    render(x: string) {
+      return html`<input class=${x} />`;
+    },
+    expectations: [
+      {
+        args: ['TEST'],
+        html: '<input class="TEST">',
+      },
+      {
+        args: ['TEST2'],
+        html: '<input class="TEST2">',
+      },
+    ],
+    stableSelectors: ['input'],
+  },
+
   /******************************************************
    * PropertyPart tests
    ******************************************************/
@@ -2792,6 +2813,27 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
+  'PropertyPart on void element': {
+    render(x: string) {
+      return html`<input .title=${x} />`;
+    },
+    expectations: [
+      {
+        args: ['TEST'],
+        html: '<input title="TEST">',
+      },
+      {
+        args: ['TEST2'],
+        html: '<input title="TEST2">',
+      },
+    ],
+    stableSelectors: ['input'],
+    // We set properties during hydration, and natively-reflecting properties
+    // will trigger a "mutation" even when set to the same value that was
+    // rendered to its attribute
+    expectMutationsDuringHydration: true,
+  },
+
   /******************************************************
    * EventPart tests
    ******************************************************/
@@ -3139,6 +3181,45 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['button'],
   },
 
+  'EventPart on a void element': {
+    render(listener: (e: Event) => void) {
+      return html`<input @click=${listener} />`;
+    },
+    expectations: [
+      {
+        args: [
+          (e: Event) => ((e.target as ClickableInput).__wasClicked = true),
+        ],
+        html: '<input>',
+        check(assert: Chai.Assert, dom: HTMLElement) {
+          const input = dom.querySelector('input')!;
+          input.click();
+          assert.strictEqual(
+            (input as ClickableInput).__wasClicked,
+            true,
+            'not clicked during first render'
+          );
+        },
+      },
+      {
+        args: [
+          (e: Event) => ((e.target as ClickableInput).__wasClicked2 = true),
+        ],
+        html: '<input>',
+        check(assert: Chai.Assert, dom: HTMLElement) {
+          const input = dom.querySelector('input')!;
+          input.click();
+          assert.strictEqual(
+            (input as ClickableInput).__wasClicked2,
+            true,
+            'not clicked during second render'
+          );
+        },
+      },
+    ],
+    stableSelectors: ['input'],
+  },
+
   /******************************************************
    * BooleanAttributePart tests
    ******************************************************/
@@ -3471,6 +3552,23 @@ export const tests: {[name: string]: SSRTest} = {
     stableSelectors: ['div'],
   },
 
+  'BooleanAttributePart on a void element': {
+    render(hide: boolean) {
+      return html` <input ?hidden=${hide} /> `;
+    },
+    expectations: [
+      {
+        args: [true],
+        html: '<input hidden>',
+      },
+      {
+        args: [false],
+        html: '<input>',
+      },
+    ],
+    stableSelectors: ['input'],
+  },
+
   /******************************************************
    * ElementPart tests
    ******************************************************/
@@ -3550,6 +3648,25 @@ export const tests: {[name: string]: SSRTest} = {
         },
       ],
       stableSelectors: ['div'],
+    };
+  },
+
+  'ElementPart on void element': () => {
+    const inputRef = createRef();
+    return {
+      render() {
+        return html` <input ${ref(inputRef)} /> `;
+      },
+      expectations: [
+        {
+          args: [],
+          html: '<input>',
+          check(assert: Chai.Assert) {
+            assert.equal(inputRef.value?.localName, 'input');
+          },
+        },
+      ],
+      stableSelectors: ['input'],
     };
   },
 
@@ -4577,6 +4694,93 @@ export const tests: {[name: string]: SSRTest} = {
         },
       ],
       stableSelectors: ['le-render-light'],
+    };
+  },
+
+  'LitElement: hydration ordering': () => {
+    const renderOrder: string[] = [];
+    return {
+      registerElements() {
+        // When defined in bottom-up order (as they will typically be based on
+        // import graph ordering), they should hydrate top-down
+        class LEOrder3 extends LitElement {
+          @property()
+          prop = 'from3';
+          render() {
+            renderOrder.push(this.localName);
+            return html`le-order3:${this.prop}`;
+          }
+        }
+        customElements.define('le-order3', LEOrder3);
+        class LEOrder2 extends LitElement {
+          @property()
+          prop = 'from2';
+          render() {
+            renderOrder.push(this.localName);
+            return html`le-order2:${this.prop}<le-order3
+                .prop=${this.prop}
+              ></le-order3>`;
+          }
+        }
+        customElements.define('le-order2', LEOrder2);
+        class LEOrder1 extends LitElement {
+          @property()
+          prop = 'from1';
+          render() {
+            renderOrder.push(this.localName);
+            return html`le-order1:${this.prop}<le-order2
+                .prop=${this.prop}
+              ></le-order2>`;
+          }
+        }
+        customElements.define('le-order1', LEOrder1);
+        class LELight extends LitElement {
+          render() {
+            renderOrder.push(this.localName);
+            return html`le-light`;
+          }
+        }
+        customElements.define('le-light', LELight);
+      },
+      render() {
+        return html`<le-order1><le-light></le-light></le-order1>`;
+      },
+      expectations: [
+        {
+          args: [],
+          async check(assert: Chai.Assert, dom: HTMLElement) {
+            const el1 = dom.querySelector('le-order1') as LitElement;
+            await el1.updateComplete;
+            const el2 = el1?.shadowRoot?.querySelector(
+              'le-order2'
+            ) as LitElement;
+            await el2.updateComplete;
+            const el3 = el2?.shadowRoot?.querySelector(
+              'le-order3'
+            ) as LitElement;
+            await el3.updateComplete;
+            assert.deepEqual(renderOrder, [
+              'le-order1',
+              'le-light',
+              'le-order2',
+              'le-order3',
+            ]);
+          },
+          html: {
+            root: `<le-order1><le-light></le-light></le-order1>`,
+            'le-order1': {
+              root: `le-order1:from1\n<le-order2></le-order2>`,
+              'le-order2': {
+                root: `le-order2:from1\n<le-order3></le-order3>`,
+                'le-order3': {
+                  root: 'le-order3:from1',
+                },
+              },
+            },
+          },
+        },
+      ],
+      stableSelectors: ['le-order1'],
     };
   },
 };