+++ packages/lit-element/CHANGELOG.md
@@ -19,6 +19,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+### Added
+
+- Added `defer-hydration` attribute handling to `experimental-hydrate-support`,
+  which helps coordinate ordered wakeup of custom elements during hydration.
+
 ### Changed
 
 - (Since 3.0.0-pre.4) Renamed `hydrate-support` module to `experimental-hydrate-support` to reflect its experimental nature. Experimental modules may undergo breaking changes within otherwise non-major releases.