+++ packages/lit-html/CHANGELOG.md
@@ -19,6 +19,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+### Added
+
+- Added `defer-hydration` attribute handling to `experimental-hydrate`, which helps
+  coordinate ordered wakeup of custom elements during hydration.
+
 ### Changed
 
 - (Since 2.0.0-pre.7) Renamed `hydrate` module to `experimental-hydrate` to reflect its experimental nature. Experimental modules may undergo breaking changes within otherwise non-major releases.